feat(cijoe/driverkit): add configs for macos/driverkit
Simple configuration for testing MacVFN and DriverKit on macOS. This
requires MacVFN is already present on the target machine, as the process
involves user interaction in the GUI.

Because of the lacking QEMU support, this test has to run on bare-metal.
It's both compatible with Intel and Apple Silicon systems.

Signed-off-by: Mads Ynddal <m.ynddal@samsung.com>